Refactor allocator code to extract inline allocation
